Allow the Handle to send arbitrary Baseview events
This change is nessecary because a client may need to pass along keyboard
events if they are captured by another program. This often happens with,
for example, VSTs whose host captures keypresses and passes it to the VST
for it to handle. The VST then needs to pass the keypress into iced for
iced to get the keypress.